Code restructurization step 1/5: split into core and tg front-end, decompose core
Split code into core and telegram front-end modules, decomposed core into submodules to be used by each other and front-end.

WARNING: this is not a completed action, just one step. The code currently has some terrible parts and is not even supposed to work correctly all together. There are other steps to be performed.
Besides, the `core.invitations` is not decomposed well enough, it is yet to be reorganised further (to be split into back-end and front-end parts, among other things), but this is probably not a subject to this patch sequence. Although, who knows