Code restructurings step 3/5: object-orient core
Updated core module files to use OOP. This is to make interaction of front-end and back-end simpler (by using methods, which associated with a particular object, instead of using functions, which are free and aren't associated with anything) and extend potential use cases (for example, with classes it would be much easier to have multiple bots running in the same application). From now on, `core` doesn't have a config file, all the data must be provided when initializing classes (there'll be more info, see #23).
At the moment the worst parts are db_connector (not going to fix it soon, see #34) and invitations, which must be split into back-end and front-end parts (not sure, how soon this is going to happen).

WARNING: this code is not working, because the front-end part was not updated yet to comply with the back-end's updates (made in this commit)