Code restructurings step 4/5: adapt telegram front-end to core
Made use of the new core (`ChatBotCore` object, more precisely) in the telegram front-end, updated references to core functions. Moved (and renamed) `conversation_starter_finisher_lock` to `ConversationsController`, wrote some docs on it. However, the way it works now may yet be a subject to change when working on #37.

WARNING: the code is not working now due to circular imports. This could be easily resolved by decomposing invitations, but I'm not sure if it's the right thing to do right now. The patch seems to be getting very big already... We'll see tomorrow